+++ server/helpers/welcome.js
@@ -38,6 +38,7 @@ const welcomeHtml = `
         .donate-address {
             font-weight: bold;
             color: #000;
+            cursor: pointer;
         }
         hr {
             background: #ececec;
@@ -48,11 +49,12 @@ const welcomeHtml = `
         .mina-version-wrapper {
             margin-top: 10px;
             border: 1px solid #ececec;
-            padding: 8px 0;
+            padding: 8px 16px;
+            background: #8185dc;
+            color: #fff;
         }
         .mina-version {
             font-weight: bold;
-            color: #000; 
         }
     </style>
     <body>
@@ -61,11 +63,11 @@ const welcomeHtml = `
             <p class="subtitle">CONVENIENT MONITORING OF YOUR MINA NODES!</p>     
             <p class="copyright">Copyright 2021 by <a href="https://pimenov.com.ua">Serhii Pimenov</a></p>
             <p class="version">Mina Monitor v%VER%</p>
+            <p class="donate">--= You can donate Mina to address =--</p>
+            <p class="donate-address" title="Click to copy to clipboard">B62qqQjC8zaU8XXaeqb9rZXFSX9x12mCgjrdCQuJbXuxU2KUPFcH7aY</p>
             
             <p class="mina-version-wrapper"><small>MINA: </small> <span class="mina-version">%MINA%</span></p>    
 
-            <p class="donate">You can donate Mina to address</p>
-            <p class="donate-address" title="Click to copy to clipboard">B62qqQjC8zaU8XXaeqb9rZXFSX9x12mCgjrdCQuJbXuxU2KUPFcH7aY</p>
         </div>
         <script>
             const donateAddress = document.querySelector(".donate-address")